backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.1",
    "fastapi>=0.125.0",
    "huggingface-hub>=0.36.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "sam3",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
sam3 = { git = "https://github.com/facebookresearch/sam3.git" }
